Update Selenium_Automation.py
As API exposed by CoWIN is caching the data and the accurate responses can be as late as 30 mins. So made an automation script which will login into the portal(you need to fill in the OTP once) and mail you the available slots.
PS:- This is only for the people who are struggling to book their first slot.